Lesson09/requirements.txt
anyio==3.6.1
APScheduler==3.9.1
autopep8==1.7.0
cachetools==5.2.0
certifi==2022.9.24
emoji==2.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
isOdd==0.1.2
numpy==1.23.3
pandas==1.5.0
progress==1.6
pycodestyle==2.9.1
pygame==2.1.2
python-dateutil==2.8.2
python-telegram-bot==20.0a4
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tic_tac_toe_miked==0.4
toml==0.10.2
tornado==6.2
tzdata==2022.4
tzlocal==4.2